Separate system contract module
